// prisma/generated/typegraphql-prisma/resolvers/inputs/VerificationRequestWhereInput.ts
import * as TypeGraphQL from "type-graphql";
import * as GraphQLScalars from "graphql-scalars";

import { DecimalJSScalar } from "../../scalars";
import { DateTimeFilter } from "../inputs/DateTimeFilter";
import { StringFilter } from "../inputs/StringFilter";

@TypeGraphQL.InputType({
  isAbstract: true
})
export class VerificationRequestWhereInput {
  @TypeGraphQL.Field(_type => [VerificationRequestWhereInput], {
    nullable: true
  })
  AND?: VerificationRequestWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [VerificationRequestWhereInput], {
    nullable: true
  })
  OR?: VerificationRequestWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [VerificationRequestWhereInput], {
    nullable: true
  })
  NOT?: VerificationRequestWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  id?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  identifier?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  token?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  expires?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  createdAt?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  updatedAt?: DateTimeFilter | undefined;
}
